# docs/swagger.yaml
basePath: /
definitions:
  api.CreateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2024
        type: string
      price:
        example: 999
        minimum: 0
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
    type: object
  api.SubscriptionResponse:
    properties:
      created_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      end_date:
        example: 12-2024
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      price:
        example: 999
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 01-2024
        type: string
      updated_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
    type: object
  api.SummaryResponse:
    properties:
      total:
        example: 1000
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: HTTP API for user subscriptions with month-granular billing windows.
    Subscription period fields use MM-YYYY in JSON; created_at/updated_at use RFC3339
    UTC.
  title: Online Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.SubscriptionResponse'
            type: array
        "500":
          description: Internal error
          schema:
            type: string
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Request body uses MM-YYYY for start_date and optional end_date.
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.SubscriptionResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubscriptionResponse'
        "404":
          description: Not found
          schema:
            type: string
      summary: Get subscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      - description: Payload (same shape as create)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubscriptionResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Computes total price × overlapping months for subscriptions in
        range; query dates are MM-YYYY.
      parameters:
      - description: Range start MM-YYYY
        in: query
        name: start
        required: true
        type: string
      - description: Range end MM-YYYY
        in: query
        name: end
        required: true
        type: string
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SummaryResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Subscription spend summary
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"